Move RA World traits, FrozenUnderFog Modifier and various other traits into Traits
