+++ OpenRA.Mods.D2k/Traits/World/D2kResourceLayer.cs
@@ -11,7 +11,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using OpenRA.Mods.RA;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.D2k.Traits
