+++ OpenRA.Mods.RA/AI/BaseBuilder.cs
@@ -11,8 +11,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.Common.Power;
+using OpenRA.Mods.RA.Buildings;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA.AI
