+++ OpenRA.Mods.RA/Activities/Attack.cs
@@ -9,6 +9,7 @@
 #endregion
 
 using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA.Activities
