+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -152,9 +152,6 @@
     <Compile Include="Attack\AttackWander.cs" />
     <Compile Include="AutoHeal.cs" />
     <Compile Include="AutoTarget.cs" />
-    <Compile Include="BaseBuilding.cs" />
-    <Compile Include="Bridge.cs" />
-    <Compile Include="BridgeLayer.cs" />
     <Compile Include="Buildable.cs" />
     <Compile Include="CaptureNotification.cs" />
     <Compile Include="Buildings\Building.cs" />
@@ -174,7 +171,6 @@
     <Compile Include="CashTrickler.cs" />
     <Compile Include="ChronoshiftPaletteEffect.cs" />
     <Compile Include="Chronoshiftable.cs" />
-    <Compile Include="Cloak.cs" />
     <Compile Include="ConquestVictoryConditions.cs" />
     <Compile Include="Crate.cs" />
     <Compile Include="CrateAction.cs" />
@@ -197,17 +193,11 @@
     <Compile Include="EmitInfantryOnSell.cs" />
     <Compile Include="EngineerRepair.cs" />
     <Compile Include="Explodes.cs" />
-    <Compile Include="Fake.cs" />
-    <Compile Include="FreeActor.cs" />
     <Compile Include="GivesBounty.cs" />
     <Compile Include="GivesExperience.cs" />
     <Compile Include="Guard.cs" />
-    <Compile Include="Harvester.cs" />
-    <Compile Include="HarvesterHuskModifier.cs" />
-    <Compile Include="Husk.cs" />
     <Compile Include="Infiltration\InfiltrateForSupportPower.cs" />
     <Compile Include="Invulnerable.cs" />
-    <Compile Include="LeavesHusk.cs" />
     <Compile Include="Captures.cs" />
     <Compile Include="LimitedAmmo.cs" />
     <Compile Include="Lint\CheckActorReferences.cs" />
@@ -217,7 +207,6 @@
     <Compile Include="MadTank.cs" />
     <Compile Include="Mine.cs" />
     <Compile Include="Minelayer.cs" />
-    <Compile Include="Modifiers\FrozenUnderFog.cs" />
     <Compile Include="Move\Drag.cs" />
     <Compile Include="Move\Mobile.cs" />
     <Compile Include="Move\Move.cs" />
@@ -226,7 +215,6 @@
     <Compile Include="Orders\PlaceBuildingOrderGenerator.cs" />
     <Compile Include="Orders\GlobalButtonOrderGenerator.cs" />
     <Compile Include="Orders\RepairOrderGenerator.cs" />
-    <Compile Include="OreRefinery.cs" />
     <Compile Include="ParaDrop.cs" />
     <Compile Include="Passenger.cs" />
     <Compile Include="Player\PlayerStatistics.cs" />
@@ -241,7 +229,6 @@
     <Compile Include="Scripting\Properties\GuardProperties.cs" />
     <Compile Include="Widgets\Logic\TabCompletionLogic.cs" />
     <Compile Include="Player\TechTree.cs" />
-    <Compile Include="PrimaryBuilding.cs" />
     <Compile Include="Production.cs" />
     <Compile Include="ProductionBar.cs" />
     <Compile Include="ProximityCaptor.cs" />
@@ -274,8 +261,6 @@
     <Compile Include="Render\WithSmoke.cs" />
     <Compile Include="Repairable.cs" />
     <Compile Include="RepairableNear.cs" />
-    <Compile Include="RepairsUnits.cs" />
-    <Compile Include="Reservable.cs" />
     <Compile Include="ScaredyCat.cs" />
     <Compile Include="SeedsResource.cs" />
     <Compile Include="SelfHealing.cs" />
@@ -295,12 +280,10 @@
     <Compile Include="SupportPowers\SupportPowerChargeBar.cs" />
     <Compile Include="SupportPowers\SupportPowerManager.cs" />
     <Compile Include="TakeCover.cs" />
-    <Compile Include="TargetableBuilding.cs" />
-    <Compile Include="TargetableSubmarine.cs" />
-    <Compile Include="TargetableUnit.cs" />
     <Compile Include="ThrowsParticle.cs" />
     <Compile Include="TraitsInterfaces.cs" />
     <Compile Include="Traits\Air\Aircraft.cs" />
+    <Compile Include="Traits\Air\AttackBomber.cs" />
     <Compile Include="Traits\Air\AttackHeli.cs" />
     <Compile Include="Traits\Air\AttackPlane.cs" />
     <Compile Include="Traits\Air\FlyAwayOnIdle.cs" />
@@ -309,6 +292,30 @@
     <Compile Include="Traits\Air\Plane.cs" />
     <Compile Include="Traits\Air\ReturnOnIdle.cs" />
     <Compile Include="Traits\Air\TargetableAircraft.cs" />
+    <Compile Include="Traits\Buildings\BaseBuilding.cs" />
+    <Compile Include="Traits\Buildings\Bridge.cs" />
+    <Compile Include="Traits\Buildings\BridgeHut.cs" />
+    <Compile Include="Traits\Buildings\Fake.cs" />
+    <Compile Include="Traits\Buildings\FreeActor.cs" />
+    <Compile Include="Traits\Buildings\OreRefinery.cs" />
+    <Compile Include="Traits\Buildings\PrimaryBuilding.cs" />
+    <Compile Include="Traits\Buildings\RepairsUnits.cs" />
+    <Compile Include="Traits\Buildings\Reservable.cs" />
+    <Compile Include="Traits\Buildings\TargetableBuilding.cs" />
+    <Compile Include="Traits\Cloak.cs" />
+    <Compile Include="Traits\Harvester.cs" />
+    <Compile Include="Traits\HarvesterHuskModifier.cs" />
+    <Compile Include="Traits\Husk.cs" />
+    <Compile Include="Traits\LeavesHusk.cs" />
+    <Compile Include="Traits\Modifiers\FrozenUnderFog.cs" />
+    <Compile Include="Traits\RepairsBridges.cs" />
+    <Compile Include="Traits\TargetableSubmarine.cs" />
+    <Compile Include="Traits\TargetableUnit.cs" />
+    <Compile Include="Traits\World\BridgeLayer.cs" />
+    <Compile Include="Traits\World\BuildableTerrainLayer.cs" />
+    <Compile Include="Traits\World\DomainIndex.cs" />
+    <Compile Include="Traits\World\PathfinderDebugOverlay.cs" />
+    <Compile Include="Traits\World\ResourceLayer.cs" />
     <Compile Include="TransformOnCapture.cs" />
     <Compile Include="TransformOnPassenger.cs" />
     <Compile Include="Transforms.cs" />
@@ -352,9 +359,7 @@
     <Compile Include="Armament.cs" />
     <Compile Include="Buildings\BaseProvider.cs" />
     <Compile Include="Widgets\Logic\ObserverShroudSelectorLogic.cs" />
-    <Compile Include="RepairsBridges.cs" />
     <Compile Include="Activities\RepairBridge.cs" />
-    <Compile Include="BridgeHut.cs" />
     <Compile Include="Lint\CheckSequences.cs" />
     <Compile Include="Widgets\Logic\SpawnSelectorTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\ClientTooltipLogic.cs" />
@@ -369,7 +374,6 @@
     <Compile Include="Render\WithResources.cs" />
     <Compile Include="Render\WithHarvestAnimation.cs" />
     <Compile Include="Widgets\Logic\SimpleTooltipLogic.cs" />
-    <Compile Include="World\DomainIndex.cs" />
     <Compile Include="Widgets\Logic\WorldTooltipLogic.cs" />
     <Compile Include="Buildings\Bib.cs" />
     <Compile Include="Render\WithIdleOverlay.cs" />
@@ -381,12 +385,9 @@
     <Compile Include="AI\States\AirStates.cs" />
     <Compile Include="Widgets\Logic\InstallLogic.cs" />
     <Compile Include="CombatDebugOverlay.cs" />
-    <Compile Include="World\PathfinderDebugOverlay.cs" />
-    <Compile Include="AttackBomber.cs" />
     <Compile Include="Render\WithCrateBody.cs" />
     <Compile Include="Widgets\Logic\GameTimerLogic.cs" />
     <Compile Include="Widgets\Logic\ReplayControlBarLogic.cs" />
-    <Compile Include="World\BuildableTerrainLayer.cs" />
     <Compile Include="Buildings\LaysTerrain.cs" />
     <Compile Include="Attack\AttackFollow.cs" />
     <Compile Include="Attack\AttackGarrisoned.cs" />
@@ -411,7 +412,6 @@
     <Compile Include="Player\ProvidesCustomPrerequisite.cs" />
     <Compile Include="Lint\CheckActors.cs" />
     <Compile Include="Lint\CheckMapCordon.cs" />
-    <Compile Include="World\ResourceLayer.cs" />
     <Compile Include="Parachutable.cs" />
     <Compile Include="Render\WithMakeAnimation.cs" />
     <Compile Include="Widgets\Logic\InstallFromCDLogic.cs" />
