+++ OpenRA.Mods.RA/Orders/PlaceBuildingOrderGenerator.cs
@@ -13,9 +13,10 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common;
-using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.Common.Graphics;
+using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.RA.Render;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 using OpenRA.Primitives;
