+++ OpenRA.Mods.RA/Traits/Buildings/RepairsUnits.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	public class RepairsUnitsInfo : TraitInfo<RepairsUnits>
 	{
