+++ OpenRA.Mods.RA/Traits/Buildings/TargetableBuilding.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	public class TargetableBuildingInfo : ITraitInfo, ITargetableInfo, Requires<BuildingInfo>
 	{
