+++ OpenRA.Mods.RA/Traits/Harvester.cs
@@ -17,7 +17,7 @@ using OpenRA.Mods.RA.Activities;
 using OpenRA.Mods.RA.Move;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	public class HarvesterInfo : ITraitInfo
 	{
