+++ OpenRA.Mods.RA/Traits/LeavesHusk.cs
@@ -14,7 +14,7 @@ using OpenRA.Mods.RA.Move;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	[Desc("Spawn another actor immediately upon death.")]
 	public class LeavesHuskInfo : ITraitInfo
