+++ OpenRA.Mods.RA/Traits/TargetableSubmarine.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	public class TargetableSubmarineInfo : TargetableUnitInfo, Requires<CloakInfo>
 	{
