+++ OpenRA.Mods.RA/Traits/TargetableUnit.cs
@@ -11,7 +11,7 @@
 using System.Collections.Generic;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	[Desc("Actor can be targeted.")]
 	public class TargetableUnitInfo : ITraitInfo, ITargetableInfo
