+++ OpenRA.Mods.RA/Traits/World/BridgeLayer.cs
@@ -14,7 +14,7 @@ using OpenRA.Graphics;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	class BridgeLayerInfo : ITraitInfo
 	{
