+++ OpenRA.Mods.RA/Widgets/Logic/ObserverStatsLogic.cs
@@ -15,6 +15,7 @@ using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.Power;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Network;
 using OpenRA.Traits;
 using OpenRA.Widgets;
